move get_reboot_status to api module
We are now concentrating every reboot_required logic
into the API. The security_status module will now
consume the reboot information directly from the API
module
