+++ uaclient/tests/test_security_status.py
@@ -5,22 +5,21 @@ import mock
 import pytest
 
 from uaclient import livepatch
+from uaclient.api.u.pro.security.status.reboot_required.v1 import RebootStatus
 from uaclient.entitlements.entitlement_status import (
     ApplicationStatus,
     ContractStatus,
 )
 from uaclient.security_status import (
-    RebootStatus,
     UpdateStatus,
     filter_security_updates,
     get_livepatch_fixed_cves,
     get_origin_for_package,
-    get_reboot_status,
     get_ua_info,
     get_update_status,
     security_status_dict,
 )
-from uaclient.system import KernelInfo, RebootRequiredPkgs
+from uaclient.system import KernelInfo
 
 M_PATH = "uaclient.security_status."
 
@@ -531,7 +530,7 @@ class TestSecurityStatus:
                 == "not-a-security-update"
             )
 
-    @mock.patch(M_PATH + "get_reboot_status")
+    @mock.patch(M_PATH + "_reboot_required")
     @mock.patch(M_PATH + "get_livepatch_fixed_cves", return_value=[])
     @mock.patch(
         M_PATH + "_is_attached", return_value=mock.MagicMock(is_attached=False)
@@ -560,7 +559,9 @@ class TestSecurityStatus:
             "esm-apps": [],
             "standard-security": [],
         }
-        m_reboot_status.return_value = RebootStatus.REBOOT_NOT_REQUIRED
+        m_reboot_status.return_value = mock.MagicMock(
+            reboot_required=RebootStatus.REBOOT_NOT_REQUIRED.value
+        )
 
         expected_output = {
             "_schema_version": "0.1",
@@ -673,190 +674,3 @@ class TestGetLivepatchFixedCVEs:
         assert [
             {"name": "cve-example", "patched": True}
         ] == get_livepatch_fixed_cves()
-
-
-class TestRebootStatus:
-    @mock.patch("uaclient.security_status.should_reboot", return_value=False)
-    def test_get_reboot_status_no_reboot_needed(self, m_should_reboot):
-        assert get_reboot_status() == RebootStatus.REBOOT_NOT_REQUIRED
-        assert 1 == m_should_reboot.call_count
-
-    @mock.patch("uaclient.security_status.get_reboot_required_pkgs")
-    @mock.patch("uaclient.security_status.should_reboot", return_value=True)
-    def test_get_reboot_status_no_reboot_pkgs_file(
-        self, m_should_reboot, m_get_reboot_required_pkgs
-    ):
-        m_get_reboot_required_pkgs.return_value = None
-        assert get_reboot_status() == RebootStatus.REBOOT_REQUIRED
-        assert 1 == m_should_reboot.call_count
-        assert 1 == m_get_reboot_required_pkgs.call_count
-
-    @mock.patch("uaclient.security_status.livepatch.status")
-    @mock.patch(
-        "uaclient.security_status.livepatch.is_livepatch_installed",
-        return_value=True,
-    )
-    @mock.patch("uaclient.security_status.get_reboot_required_pkgs")
-    @mock.patch("uaclient.security_status.should_reboot", return_value=True)
-    def test_get_reboot_status_livepatch_status_none(
-        self,
-        m_should_reboot,
-        m_get_reboot_required_pkgs,
-        _m_is_livepatch_installed,
-        m_livepatch_status,
-    ):
-        m_get_reboot_required_pkgs.return_value = RebootRequiredPkgs(
-            standard_packages=[],
-            kernel_packages=["linux-base", "linux-image-5.4.0-1074"],
-        )
-        m_livepatch_status.return_value = None
-        assert get_reboot_status() == RebootStatus.REBOOT_REQUIRED
-
-    @pytest.mark.parametrize(
-        "reboot_required_pkgs,expected_state",
-        (
-            (
-                RebootRequiredPkgs(
-                    standard_packages=["pkg1", "pkg2"], kernel_packages=[]
-                ),
-                RebootStatus.REBOOT_REQUIRED,
-            ),
-            (
-                RebootRequiredPkgs(
-                    standard_packages=["pkg2"],
-                    kernel_packages=["linux-base", "linux-image-5.4.0-1074"],
-                ),
-                RebootStatus.REBOOT_REQUIRED,
-            ),
-        ),
-    )
-    @mock.patch("uaclient.security_status.get_reboot_required_pkgs")
-    @mock.patch("uaclient.security_status.should_reboot", return_value=True)
-    def test_get_reboot_status_reboot_pkgs_file_present(
-        self,
-        m_should_reboot,
-        m_get_reboot_required_pkgs,
-        reboot_required_pkgs,
-        expected_state,
-    ):
-        m_get_reboot_required_pkgs.return_value = reboot_required_pkgs
-        assert get_reboot_status() == expected_state
-        assert 1 == m_should_reboot.call_count
-        assert 1 == m_get_reboot_required_pkgs.call_count
-
-    @pytest.mark.parametrize(
-        [
-            "livepatch_state",
-            "supported_state",
-            "expected_state",
-            "kernel_name",
-        ],
-        (
-            (
-                "applied",
-                "supported",
-                RebootStatus.REBOOT_REQUIRED_LIVEPATCH_APPLIED,
-                "4.15.0-187.198-generic",
-            ),
-            (
-                "applied",
-                None,
-                RebootStatus.REBOOT_REQUIRED,
-                "4.15.0-187.198-generic",
-            ),
-            ("applied", "supported", RebootStatus.REBOOT_REQUIRED, "test"),
-            (
-                "nothing-to-apply",
-                "supported",
-                RebootStatus.REBOOT_REQUIRED_LIVEPATCH_APPLIED,
-                "4.15.0-187.198-generic",
-            ),
-            (
-                "applying",
-                "supported",
-                RebootStatus.REBOOT_REQUIRED,
-                "4.15.0-187.198-generic",
-            ),
-            (
-                "apply-failed",
-                "supported",
-                RebootStatus.REBOOT_REQUIRED,
-                "4.15.0-187.198-generic",
-            ),
-        ),
-    )
-    @mock.patch("uaclient.security_status.get_kernel_info")
-    @mock.patch("uaclient.security_status.livepatch.is_livepatch_installed")
-    @mock.patch("uaclient.security_status.livepatch.status")
-    @mock.patch("uaclient.security_status.get_reboot_required_pkgs")
-    @mock.patch("uaclient.security_status.should_reboot", return_value=True)
-    def test_get_reboot_status_reboot_pkgs_file_only_kernel_pkgs(
-        self,
-        m_should_reboot,
-        m_get_reboot_required_pkgs,
-        m_livepatch_status,
-        m_is_livepatch_installed,
-        m_kernel_info,
-        livepatch_state,
-        supported_state,
-        expected_state,
-        kernel_name,
-    ):
-        m_kernel_info.return_value = mock.MagicMock(
-            proc_version_signature_version=kernel_name
-        )
-        m_is_livepatch_installed.return_value = True
-        m_get_reboot_required_pkgs.return_value = RebootRequiredPkgs(
-            standard_packages=[],
-            kernel_packages=["linux-base", "linux-image-5.4.0-1074"],
-        )
-        m_livepatch_status.return_value = livepatch.LivepatchStatusStatus(
-            kernel="4.15.0-187.198-generic",
-            livepatch=livepatch.LivepatchPatchStatus(
-                state=livepatch_state, fixes=None, version=None
-            ),
-            supported=supported_state,
-        )
-
-        assert get_reboot_status() == expected_state
-        assert 1 == m_should_reboot.call_count
-        assert 1 == m_get_reboot_required_pkgs.call_count
-        assert 1 == m_is_livepatch_installed.call_count
-        assert 1 == m_livepatch_status.call_count
-        assert 1 == m_kernel_info.call_count
-
-    @mock.patch("uaclient.security_status.get_kernel_info")
-    @mock.patch("uaclient.security_status.livepatch.is_livepatch_installed")
-    @mock.patch("uaclient.security_status.livepatch.status")
-    @mock.patch("uaclient.security_status.get_reboot_required_pkgs")
-    @mock.patch("uaclient.security_status.should_reboot", return_value=True)
-    def test_get_reboot_status_fail_parsing_kernel_info(
-        self,
-        m_should_reboot,
-        m_get_reboot_required_pkgs,
-        m_livepatch_status,
-        m_is_livepatch_installed,
-        m_kernel_info,
-    ):
-        m_kernel_info.return_value = mock.MagicMock(
-            proc_version_signature_version=None
-        )
-        m_is_livepatch_installed.return_value = True
-        m_get_reboot_required_pkgs.return_value = RebootRequiredPkgs(
-            standard_packages=[],
-            kernel_packages=["linux-base", "linux-image-5.4.0-1074"],
-        )
-        m_livepatch_status.return_value = livepatch.LivepatchStatusStatus(
-            kernel="4.15.0-187.198-generic",
-            livepatch=livepatch.LivepatchPatchStatus(
-                state="applied", fixes=None, version=None
-            ),
-            supported=None,
-        )
-
-        assert get_reboot_status() == RebootStatus.REBOOT_REQUIRED
-        assert 1 == m_should_reboot.call_count
-        assert 1 == m_get_reboot_required_pkgs.call_count
-        assert 1 == m_is_livepatch_installed.call_count
-        assert 1 == m_livepatch_status.call_count
-        assert 1 == m_kernel_info.call_count
